Split devices.js into keyboard.js and mouse.js
