+++ tests/input.html
@@ -31,7 +31,8 @@
     <script src="../core/input/xtscancodes.js"></script>
     <script src="../core/input/vkeys.js"></script>
     <script src="../core/input/util.js"></script>
-    <script src="../core/input/devices.js"></script>
+    <script src="../core/input/keyboard.js"></script>
+    <script src="../core/input/mouse.js"></script>
     <script src="../core/display.js"></script>
     <script>
         var msg_cnt = 0, iterations,
