+++ tests/vnc_perf.html
@@ -50,8 +50,8 @@
             WebUtil.load_scripts({
                 'core': ["base64.js", "websock.js", "des.js", "input/keysym.js",
                          "input/keysymdef.js", "input/xtscancodes.js", "input/util.js",
-                         "input/devices.js", "display.js", "rfb.js", "inflator.js",
-                         "input/vkeys.js", "input/fixedkeys.js"],
+                         "input/keyboard.js", "input/mouse.js", "display.js", "rfb.js",
+                         "inflator.js", "input/vkeys.js", "input/fixedkeys.js"],
                 'tests': ["playback.js"],
                 'recordings': [fname]});
         } else {
